Extract key pair to its own file, rename and conform to Codable
