+++ src/xcode/ENA/ENA/Source/Client/Security/iOS 12 Fallback/CryptoKitFallbacks.swift
@@ -346,68 +346,3 @@ extension String {
 		}
 	}
 }
-
-
-// MARK: - Digital Green Certificate
-
-struct DGCRSAKeypair {
-	
-	enum DGCRSAKeypairError: Error {
-		case keypairGeneration(String)	// Key generation failed
-		case publicKey // Unable to get public key representation
-	}
-	
-	let publicKey: SecKey
-	let privateKey: SecKey
-	
-	/// The publicKey with added RSA Header and Base64 encoded
-	let publicKeyForBackend: String
-	
-	init() throws {
-		let tag = Bundle.main.bundleIdentifier ?? "de.rki.coronawarnapp"
-		
-		let publicKeyAttr: [NSObject: NSObject] = [
-			kSecAttrIsPermanent: true as NSObject,
-			kSecAttrApplicationTag: ("\(tag).dgc.public".data(using: String.Encoding.utf8) ?? Data()) as NSObject,
-			kSecClass: kSecClassKey,
-			kSecReturnData: kCFBooleanTrue]
-		
-		let privateKeyAttr: [NSObject: NSObject] = [
-			kSecAttrIsPermanent: true as NSObject,
-			kSecAttrApplicationTag: ("\(tag).dgc.private".data(using: String.Encoding.utf8) ?? Data()) as NSObject,
-			kSecClass: kSecClassKey,
-			kSecReturnData: kCFBooleanTrue]
-		
-		var keyPairAttr = [NSObject: NSObject]()
-		keyPairAttr[kSecAttrKeyType] = kSecAttrKeyTypeRSA
-		keyPairAttr[kSecAttrKeySizeInBits] = 3072 as NSObject
-		keyPairAttr[kSecPublicKeyAttrs] = publicKeyAttr as NSObject
-		keyPairAttr[kSecPrivateKeyAttrs] = privateKeyAttr as NSObject
-		
-		var _publicKey: SecKey?
-		var _privateKey: SecKey?
-		
-		let statusCode = SecKeyGeneratePair(keyPairAttr as CFDictionary, &_publicKey, &_privateKey)
-		
-		guard statusCode == noErr, let publicKey = _publicKey, let privateKey = _privateKey else {
-			let errorText = String(describing: statusCode)
-			Log.error("Error generating DGC RSA key pair: \(errorText)", log: .crypto)
-			throw DGCRSAKeypairError.keypairGeneration(errorText)
-		}
-		self.privateKey = privateKey
-		self.publicKey = publicKey
-		
-		var error: Unmanaged<CFError>?
-		guard let publicKeyData = SecKeyCopyExternalRepresentation(publicKey, &error) as Data? else {
-			throw DGCRSAKeypairError.publicKey
-		}
-		let publicKeyWithRSAHeader = Data([
-			0x30, 0x82, 0x01, 0xA2,
-			0x30, 0x0D,
-			0x06, 0x09, 0x2A, 0x86, 0x48, 0x86, 0xF7, 0x0D, 0x01, 0x01, 0x01,
-			0x05, 0x00,
-			0x03, 0x82, 0x01, 0x8F, 0x00
-		]) + publicKeyData
-		self.publicKeyForBackend = publicKeyWithRSAHeader.base64EncodedString()
-	}
-}
